job_project/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-filter==23.3
django-phonenumber-field==7.3.0
django-widget-tweaks==1.5.0
phonenumbers==8.13.27
sqlparse==0.4.4
tzdata==2023.3
